pyproject.toml
[tool.poetry]
name = "epic_events_crm"
version = "0.1.0"
description = "School project - Simple CRM as CLI"
authors = ["DayriseA"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.27"
alembic = "^1.13.1"
click = "^8.1.7"
rich = "^13.7.0"
pymysql = "^1.1.0"
cryptography = "^42.0.4"
python-dotenv = "^1.0.1"
argon2-cffi = "^23.1.0"
pyjwt = "^2.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
